[Feat] TestViewModel 통합 및 LiveData 삭제
